requirements.txt
fastapi[standard]
docxtpl
python-dotenv
docx2pdf
python-docx